build: move MetadataIconData.kt and AvatarProviderConfiguration.kt to commonsKmm
