+++ app/src/main/java/org/dhis2/usescases/development/RuleValidation.kt
@@ -1,6 +1,6 @@
 package org.dhis2.usescases.development
 
-import org.dhis2.ui.MetadataIconData
+import org.dhis2.mobile.commons.model.MetadataIconData
 import org.hisp.dhis.rules.models.Rule
 
 data class ProgramRuleValidation(
