+++ app/src/main/java/org/dhis2/usescases/main/program/ProgramUi.kt
@@ -58,7 +58,7 @@ import org.dhis2.commons.date.toDateSpan
 import org.dhis2.commons.ui.icons.toIconData
 import org.dhis2.data.service.SyncStatusData
 import org.dhis2.mobile.commons.extensions.toColorInt
-import org.dhis2.ui.MetadataIconData
+import org.dhis2.mobile.commons.model.MetadataIconData
 import org.hisp.dhis.android.core.common.State
 import org.hisp.dhis.mobile.ui.designsystem.component.AdditionalInfoItem
 import org.hisp.dhis.mobile.ui.designsystem.component.Avatar
