+++ app/src/main/java/org/dhis2/usescases/programStageSelection/ProgramStageData.kt
@@ -1,6 +1,6 @@
 package org.dhis2.usescases.programStageSelection
 
-import org.dhis2.ui.MetadataIconData
+import org.dhis2.mobile.commons.model.MetadataIconData
 import org.hisp.dhis.android.core.program.ProgramStage
 
 data class ProgramStageData(
