+++ app/src/main/java/org/dhis2/usescases/teiDashboard/DashboardProgramModel.kt
@@ -1,6 +1,6 @@
 package org.dhis2.usescases.teiDashboard
 
-import org.dhis2.ui.MetadataIconData
+import org.dhis2.mobile.commons.model.MetadataIconData
 import org.hisp.dhis.android.core.enrollment.Enrollment
 import org.hisp.dhis.android.core.enrollment.EnrollmentStatus
 import org.hisp.dhis.android.core.organisationunit.OrganisationUnit
