+++ app/src/main/java/org/dhis2/usescases/teiDashboard/DashboardRepository.kt
@@ -3,7 +3,7 @@ package org.dhis2.usescases.teiDashboard
 import io.reactivex.Flowable
 import io.reactivex.Observable
 import io.reactivex.Single
-import org.dhis2.ui.MetadataIconData
+import org.dhis2.mobile.commons.model.MetadataIconData
 import org.hisp.dhis.android.core.category.CategoryOptionCombo
 import org.hisp.dhis.android.core.enrollment.Enrollment
 import org.hisp.dhis.android.core.enrollment.EnrollmentStatus
