+++ app/src/main/java/org/dhis2/usescases/teiDashboard/dashboardfragments/teidata/teievents/EventViewHolder.java
@@ -18,7 +18,7 @@ import org.dhis2.commons.data.EventViewModel;
 import org.dhis2.commons.databinding.ItemFieldValueBinding;
 import org.dhis2.commons.resources.ColorUtils;
 import org.dhis2.databinding.ItemEventBinding;
-import org.dhis2.ui.MetadataIconData;
+import org.dhis2.mobile.commons.model.MetadataIconData;
 import org.dhis2.utils.DhisTextUtils;
 import org.hisp.dhis.android.core.enrollment.Enrollment;
 import org.hisp.dhis.android.core.event.Event;
