+++ app/src/main/res/layout/item_dashboard_program.xml
@@ -21,7 +21,7 @@
 
         <variable
             name="metadataIconData"
-            type="org.dhis2.ui.MetadataIconData" />
+            type="org.dhis2.mobile.commons.model.MetadataIconData" />
 
         <variable
             name="orgUnit"
