+++ app/src/test/java/org/dhis2/uicomponents/map/geometry/MapTeiEventsToFeatureCollectionTest.kt
@@ -7,12 +7,11 @@ import org.dhis2.maps.geometry.point.MapPointToFeature
 import org.dhis2.maps.geometry.polygon.MapPolygonToFeature
 import org.dhis2.maps.model.MapItemModel
 import org.dhis2.maps.model.RelatedInfo
-import org.dhis2.ui.avatar.AvatarProviderConfiguration
+import org.dhis2.mobile.commons.model.AvatarProviderConfiguration
 import org.dhis2.uicomponents.map.mocks.GeometryDummy.getGeometryAsPoint
 import org.hamcrest.CoreMatchers.`is`
 import org.hamcrest.MatcherAssert.assertThat
 import org.hisp.dhis.android.core.common.State
-import org.hisp.dhis.android.core.event.Event
 import org.junit.Before
 import org.junit.Test
 import org.maplibre.geojson.Point
