+++ app/src/test/java/org/dhis2/usescases/programEventDetail/ProgramEventMapperTest.kt
@@ -3,7 +3,7 @@ package org.dhis2.usescases.programEventDetail
 import androidx.compose.ui.graphics.Color
 import org.dhis2.commons.resources.DhisPeriodUtils
 import org.dhis2.commons.resources.MetadataIconProvider
-import org.dhis2.ui.MetadataIconData
+import org.dhis2.mobile.commons.model.MetadataIconData
 import org.hisp.dhis.android.core.D2
 import org.hisp.dhis.android.core.category.CategoryOptionCombo
 import org.hisp.dhis.android.core.common.ObjectStyle
