+++ app/src/test/java/org/dhis2/usescases/tracker/TrackedEntityInstanceInfoProviderTests.kt
@@ -3,9 +3,9 @@ package org.dhis2.usescases.tracker
 import androidx.compose.ui.graphics.Color
 import org.dhis2.commons.date.DateLabelProvider
 import org.dhis2.commons.resources.MetadataIconProvider
+import org.dhis2.mobile.commons.model.AvatarProviderConfiguration
+import org.dhis2.mobile.commons.model.MetadataIconData
 import org.dhis2.tracker.data.ProfilePictureProvider
-import org.dhis2.ui.MetadataIconData
-import org.dhis2.ui.avatar.AvatarProviderConfiguration
 import org.hisp.dhis.android.core.D2
 import org.hisp.dhis.android.core.common.ObjectStyle
 import org.hisp.dhis.android.core.program.Program
