+++ commons/src/main/java/org/dhis2/commons/data/EnrollmentIconData.kt
@@ -1,6 +1,6 @@
 package org.dhis2.commons.data
 
-import org.dhis2.ui.MetadataIconData
+import org.dhis2.mobile.commons.model.MetadataIconData
 
 data class EnrollmentIconData(
     val color: Int,
