+++ commons/src/main/java/org/dhis2/commons/data/EventViewModel.kt
@@ -1,6 +1,6 @@
 package org.dhis2.commons.data
 
-import org.dhis2.ui.MetadataIconData
+import org.dhis2.mobile.commons.model.MetadataIconData
 import org.hisp.dhis.android.core.event.Event
 import org.hisp.dhis.android.core.program.ProgramStage
 import java.util.Date
