+++ form/src/main/java/org/dhis2/form/model/OptionSetConfiguration.kt
@@ -5,7 +5,7 @@ import androidx.paging.map
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.map
-import org.dhis2.ui.MetadataIconData
+import org.dhis2.mobile.commons.model.MetadataIconData
 import org.hisp.dhis.android.core.option.Option
 
 data class OptionSetConfiguration(
